added total revenue and revenue per order, home page, and about page

+++ backend/dto/customer.go
@@ -6,7 +6,7 @@ type CustomerCityCount struct {
 }
 
 type Count struct {
-	Count int `json:"count"`
+	Count float64 `json:"count"`
 }
 
 type Rate struct {
